Reorganizing the folders / implementiong new build mechanisms Updating the .gitignore file and removing unwanted folders from repo
Re-organizing code folders and implementing gradle build mechanism

Implementing quality gate for code implementations
