+++ checkstyle/rlearn-checkstyle.xml
@@ -1,0 +1,124 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE module PUBLIC "-//Puppy Crawl//DTD Check Configuration 1.3//EN" "http://www.puppycrawl.com/dtds/configuration_1_3.dtd">
+
+<!--
+    This configuration file was written by the eclipse-cs plugin configuration editor
+-->
+<!--
+    Checkstyle-Configuration: Editable RLearning
+    Description: none
+-->
+<module name="Checker">
+    <property name="severity" value="warning"/>
+    <module name="TreeWalker">
+        <module name="JavadocMethod">
+            <property name="severity" value="warning"/>
+            <property name="scope" value="protected"/>
+            <property name="excludeScope" value="package"/>
+            <property name="allowUndeclaredRTE" value="true"/>
+            <property name="allowThrowsTagsForSubclasses" value="true"/>
+            <property name="allowMissingParamTags" value="true"/>
+            <property name="allowMissingReturnTag" value="true"/>
+            <property name="allowMissingPropertyJavadoc" value="true"/>
+            <property name="logLoadErrors" value="true"/>
+            <property name="suppressLoadErrors" value="true"/>
+        </module>
+        <module name="JavadocType">
+            <property name="severity" value="warning"/>
+        </module>
+        <module name="JavadocStyle">
+            <property name="severity" value="warning"/>
+            <property name="checkFirstSentence" value="false"/>
+            <property name="checkHtml" value="false"/>
+            <property name="tokens" value="INTERFACE_DEF,CLASS_DEF,METHOD_DEF,CTOR_DEF"/>
+        </module>
+        <module name="ConstantName">
+            <property name="format" value="^([A-Z][A-Z0-9]*(_[A-Z0-9]+)*|log|logger)$"/>
+        </module>
+        <module name="LocalFinalVariableName"/>
+        <module name="LocalVariableName"/>
+        <module name="MemberName"/>
+        <module name="MethodName"/>
+        <module name="PackageName"/>
+        <module name="ParameterName"/>
+        <module name="StaticVariableName"/>
+        <module name="TypeName"/>
+        <module name="AvoidStarImport">
+            <property name="severity" value="warning"/>
+        </module>
+        <module name="IllegalImport"/>
+        <module name="RedundantImport"/>
+        <module name="UnusedImports">
+            <property name="severity" value="warning"/>
+        </module>
+        <module name="LineLength">
+            <property name="severity" value="warning"/>
+            <property name="max" value="300"/>
+            <property name="tabWidth" value="4"/>
+        </module>
+        <module name="ModifierOrder"/>
+        <module name="AvoidNestedBlocks"/>
+        <module name="NeedBraces">
+            <property name="severity" value="warning"/>
+        </module>
+        <module name="EmptyStatement"/>
+        <module name="EqualsHashCode"/>
+        <module name="HiddenField">
+            <property name="severity" value="warning"/>
+            <property name="ignoreConstructorParameter" value="true"/>
+            <property name="ignoreSetter" value="true"/>
+            <property name="ignoreAbstractMethods" value="true"/>
+        </module>
+        <module name="InnerAssignment"/>
+        <module name="RedundantThrows">
+            <property name="logLoadErrors" value="true"/>
+            <property name="suppressLoadErrors" value="true"/>
+        </module>
+        <module name="SimplifyBooleanExpression">
+            <property name="severity" value="warning"/>
+        </module>
+        <module name="SimplifyBooleanReturn">
+            <property name="severity" value="warning"/>
+        </module>
+        <module name="UpperEll">
+            <property name="severity" value="warning"/>
+        </module>
+        <module name="MultipleVariableDeclarations"/>
+        <module name="ParameterAssignment"/>
+        <module name="StringLiteralEquality"/>
+        <module name="UnnecessaryParentheses">
+            <property name="severity" value="warning"/>
+        </module>
+        <module name="ModifierOrder"/>
+        <module name="EmptyBlock"/>
+        <module name="ExplicitInitialization"/>
+        <module name="FallThrough"/>
+        <module name="MultipleStringLiterals">
+            <property name="allowedDuplicates" value="5"/>
+            <property name="ignoreStringsRegexp" value="^&quot;\s?(|\$|%|\(|\)|-|=|\{|\}|;|:|'|,|\.|/|\?|\\|\||\\n)\s?&quot;$"/>
+        </module>
+        <module name="Regexp">
+            <metadata name="com.atlassw.tools.eclipse.checkstyle.comment" value="Stop it, Roy.  Enter a useful comment."/>
+            <property name="format" value="^\/\*[\*]+\*/$"/>
+            <property name="message" value="Stop it, Roy.  Enter a useful comment."/>
+            <property name="illegalPattern" value="true"/>
+        </module>
+        <module name="Regexp">
+            <metadata name="com.atlassw.tools.eclipse.checkstyle.comment" value="Invalid use of System.out.println"/>
+            <property name="format" value="^.*System\.out\.print.*$"/>
+            <property name="message" value="Use a logging utility such as log4j"/>
+            <property name="illegalPattern" value="true"/>
+        </module>
+        <module name="Regexp">
+            <metadata name="com.atlassw.tools.eclipse.checkstyle.comment" value="Invalid use of java.util.logging"/>
+            <property name="format" value="^.*java\.util\.logging.*$"/>
+            <property name="message" value="Use a logging utility such as log4j"/>
+            <property name="illegalPattern" value="true"/>
+        </module>
+        <module name="EqualsAvoidNull"/>
+    </module>
+    <module name="StrictDuplicateCode">
+        <property name="severity" value="warning"/>
+        <property name="min" value="30"/>
+    </module>
+</module>
