+++ src/main/java/com/rlearning/controller/SuperPixelEditorController.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Controller;
+package com.rlearning.controller;
 
 /**
  *
