+++ src/main/java/com/rlearning/model/SaveColorArray.java
@@ -1,4 +1,4 @@
-package com.habim.ImageCanvas;
+package com.rlearning.model;
 
 import javafx.scene.paint.Color;
 import java.io.BufferedWriter;
