+++ src/main/java/com/rlearning/model/SuperPixelEditorModel.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Model;
+package com.rlearning.model;
 
 /**
  * @author Andrew
