+++ version.gradle
@@ -1,0 +1,3 @@
+gradle.ext {
+    javafxGradlePluginVersion = '8.8.2'
+}
